+++ PS7-x64/Microsoft.PowerShell_profile.ps1
@@ -1,7 +1,6 @@
 ﻿$ErrorActionPreference = 'SilentlyContinue'
 
-#Powershell Customization
-#Need newer version of PSReadlin that fixes tab complete bug
+<#old prompt
 function prompt {
     $location = Get-Location
     Write-Host -NoNewline "$(HOSTNAME.EXE) "                  -ForegroundColor Green
@@ -11,6 +10,21 @@ function prompt {
 
     $Adminp = [bool](([System.Security.Principal.WindowsIdentity]::GetCurrent()).groups -match "S-1-5-32-544")
     $host.UI.RawUI.WindowTitle = 'Admin is ' + "$Adminp" + ' - PSVersion ' + $host.version + " - $location"
+
+    Return " "
+}
+#>
+
+function prompt {
+    $location = Get-Location
+    Write-Host -NoNewline "$(HOSTNAME.EXE) "                  -ForegroundColor Green
+    Write-Host -NoNewline '~'                                 -ForegroundColor Yellow
+    Write-Host -NoNewline $(Get-Location).Path.Split('\')[-1] -ForegroundColor Cyan
+    Write-Host -NoNewline ">" -ForegroundColor Green
+
+    $Adminp = [bool](([System.Security.Principal.WindowsIdentity]::GetCurrent()).groups -match "S-1-5-32-544")
+    $ver = [string]$host.Version.major + '.' + [string]$host.version.minor
+    $host.UI.RawUI.WindowTitle = "$ver" + ' - Admin is ' + "$Adminp" + " - $location"
 
     Return " "
 }