+++ PS7-x64/Modules/AdminToolbox.Exchange/ChangeLog.md
@@ -12,3 +12,4 @@
 * **1.2.0.1** Made fixes to commands Grant-MailboxAccessOWAOnly and Get-MailboxScalereport
 * **1.2.0.2** Made fixes to commands Grant-MailboxAccessOWAOnly
 * **1.3.0.0** Added Enable-365MailboxExistingUser function
+* **1.4.0.0** Added Connect-ExchangeServer and disconnect-exchangeserver functions